+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # CHANGELOG
 
+## UNRELEASED
+
+- Changed routing number validity from checks against static list to dynamic checking against [standard ABA RTN formats](https://en.wikipedia.org/wiki/Routing_transit_number#MICR_routing_number_format)
+
 ## 1.0.1
 
 - Updates (sub-)dependencies
   - `braces` to 3.0.3
-  - `cross-spawn` to 7.0.6 
+  - `cross-spawn` to 7.0.6
   - `micromatch` to 4.0.8
 
 ## 1.0.0